# Conflicts:
#	src/main/java/com/javamentor/qa/platform/dao/abstracts/dto/TagDtoDao.java

+++ src/main/java/com/javamentor/qa/platform/service/impl/dto/TagDtoServiceImpl.java
@@ -32,7 +32,7 @@ public class TagDtoServiceImpl implements TagDtoService {
        pageDto.setTotalResultCount(totalResultCount);
        pageDto.setCurrentPageNumber(page);
        pageDto.setItemsOnPage(size);
-       pageDto.setTotalPageCount((int) Math.ceil(totalResultCount/size));
+       pageDto.setTotalPageCount((int) Math.ceil(totalResultCount/(double)size));
 
         return pageDto;
     }
